feat: Implement Chair Dashboard with blocking alerts and queue management
- Added a new dashboard view for Chairs, replacing the generic approver screen.
- Introduced `ProvidesDashboardAlerts` interface for modules to declare blocking conditions.
- Implemented alerts for missing signatures on hardbound submissions.
- Created various dashboard panels including queues, longest waits, and filed nominations.
- Enhanced CSS for the Chair dashboard layout and alerts.
- Updated tests to cover new dashboard functionality and alerts.

+++ app/Modules/Core/Resources/views/partials/queue.blade.php
@@ -115,7 +115,9 @@
 
         @foreach ($applications as $application)
             @php
-                $waiting = $application->submitted_at?->diffInDays() ?? 0;
+                // (int), because Carbon 3 returns a FLOAT here -- 41.000000002049
+                // for a row submitted 41 days ago, which renders verbatim.
+                $waiting = (int) ($application->submitted_at?->diffInDays() ?? 0);
                 $tone = match (true) {
                     $waiting >= $criticalAfter => 'critical',
                     $waiting >= $overdueAfter => 'warn',
